Only set a deterministic `SECRET_KEY` when in `TESTING` mode
- Do this with a validator in new feature flags module, moved out of main
- Add a minimum Shannon entropy for they secret key in production
- Reorder test imports to avoid failing config load
- Fix info endpoint tests locally
- Update CHANGELOG
