+++ pydatalab/tests/server/test_info_and_health.py
@@ -1,10 +1,8 @@
-import os
-
 import pytest
 
 
 @pytest.mark.parametrize("url_prefix", ["", "/v0", "/v0.1", "/v0.1.0"])
-def test_info_endpoint(client, url_prefix):
+def test_info_endpoint(client, url_prefix, app):
     response = client.get(f"{url_prefix}/info")
     assert response.status_code == 200
     assert all(k in response.json for k in ("data", "meta", "links"))
@@ -13,14 +11,14 @@ def test_info_endpoint(client, url_prefix):
     assert (features := attributes.get("features"))
     assert (auth := features.get("auth_mechanisms"))
     assert auth["github"] is bool(
-        os.environ.get("GITHUB_OAUTH_CLIENT_ID", None)
-        and os.environ.get("GITHUB_OAUTH_CLIENT_SECRET", None)
+        app.config.get("GITHUB_OAUTH_CLIENT_ID", None)
+        and app.config.get("GITHUB_OAUTH_CLIENT_SECRET", None)
     )
     assert auth["orcid"] is bool(
-        os.environ.get("ORCID_OAUTH_CLIENT_ID", None)
-        and os.environ.get("ORCID_OAUTH_CLIENT_SECRET", None)
+        app.config.get("ORCID_OAUTH_CLIENT_ID", None)
+        and app.config.get("ORCID_OAUTH_CLIENT_SECRET", None)
     )
-    assert auth["email"] is bool(os.environ.get("MAIL_PASSWORD", None))
+    assert auth["email"] is bool(app.config.get("MAIL_PASSWORD", None))
 
 
 @pytest.mark.parametrize("url_prefix", ["", "/v0", "/v0.1", "/v0.1.0"])
