+++ pydatalab/tests/test_config.py
@@ -2,11 +2,10 @@ from pathlib import Path
 
 import pytest
 
-from pydatalab.config import ServerConfig, SMTPSettings
-from pydatalab.main import create_app
-
 
 def test_default_settings():
+    from pydatalab.config import ServerConfig
+
     config = ServerConfig()
     assert config.MONGO_URI == "mongodb://localhost:27017/datalabvue"
     assert config.SECRET_KEY
@@ -14,6 +13,8 @@ def test_default_settings():
 
 
 def test_update_settings():
+    from pydatalab.config import ServerConfig
+
     config = ServerConfig()
     new_settings = {
         "mongo_uri": "mongodb://test",
@@ -28,6 +29,8 @@ def test_update_settings():
 
 
 def test_config_override():
+    from pydatalab.main import create_app
+
     app = create_app(
         config_override={"REMOTE_FILESYSTEMS": [{"hostname": None, "path": "/", "name": "local"}]}
     )
@@ -41,6 +44,8 @@ def test_config_override():
 
 
 def test_validators():
+    from pydatalab.config import ServerConfig
+
     # check bad prefix
     with pytest.raises(
         RuntimeError, match="Identifier prefix must be less than 12 characters long,"
@@ -54,7 +59,8 @@ def test_mail_settings_combinations(tmpdir):
     overrides can be provided as environment variables.
     """
 
-    from pydatalab.config import CONFIG
+    from pydatalab.config import CONFIG, SMTPSettings
+    from pydatalab.main import create_app
 
     CONFIG.update(
         {
@@ -82,3 +88,15 @@ def test_mail_settings_combinations(tmpdir):
     app = create_app(env_file=env_file)
     assert app.config["MAIL_PASSWORD"] == "password"  # noqa: S105
     assert app.config["MAIL_DEFAULT_SENDER"] == "test2@example.com"
+
+
+def test_key_strength_checker():
+    from pydatalab.feature_flags import _check_key_strength
+
+    with pytest.raises(RuntimeError, match="Shannon entropy"):
+        assert _check_key_strength("a" * 32)
+
+    with pytest.raises(RuntimeError, match="Shannon entropy"):
+        assert _check_key_strength("ab" * 16)
+
+    assert _check_key_strength("abcdefghijklmnopqrstuvwxyz") is None
